testing

+++ src/moe/MoEVRT/deepspeed/train_moe.py
@@ -46,16 +46,16 @@ def evaluate(rank, world_size, epoch, model_engine, logger, val_dl, loss_fn, met
     val_loss, val_metrics = 0, {k: 0 for k in cfg.train.metric.metrics}
 
     for i, data in enumerate(val_dl):
-        lr, hr = data[0].half().to(model_engine.local_rank), \
-            data[1].half().to(model_engine.local_rank)
+        lr, hr = data[0].to(model_engine.local_rank), \
+            data[1].to(model_engine.local_rank)
 
-        sr, lq = model_engine(lr)
+        sr, lq = model_engine(lr.half())
         loss = compute_loss(loss_fn, sr, hr)
 
         dist.reduce(loss, dst=0, op=dist.ReduceOp.SUM)
 
         val_loss += loss.detach().item() / world_size
-        val_metrics = running_metrics(val_metrics, metric, sr.float(), hr.float())
+        val_metrics = running_metrics(val_metrics, metric, sr.float(), hr)
 
     save_checkpoint_ds(cfg, model_engine, logger, rank)
 
@@ -63,7 +63,6 @@ def evaluate(rank, world_size, epoch, model_engine, logger, val_dl, loss_fn, met
         logger.log_dict({"Loss": val_loss / len(val_dl)}, epoch, "Val")
         logger.log_dict({k: v / len(val_dl) for k, v in val_metrics.items()}, epoch, "Val")
         logger.log_images("Val", epoch, lr.float(), sr.float(), hr.float(), lq.float())
-
 
 def run(cfg: omegaconf.DictConfig, args):
     seed_index_everything(cfg.train)
@@ -104,21 +103,21 @@ def run(cfg: omegaconf.DictConfig, args):
         train_loss, train_metrics = 0.0, {k: 0 for k in cfg.train.metric.metrics}
 
         for i, data in enumerate(train_dl):
-            lr, hr = data[0].half().to(device), data[1].half().to(device)
+            lr, hr = data[0].to(device), data[1].to(device)
 
-            sr, lq = model_engine(lr)
+            sr, lq = model_engine(lr.half())
             loss = compute_loss(loss_fn, sr, hr, lq)
 
             model_engine.backward(loss)
             model_engine.step()
 
             train_loss += loss.detach().item()
-            train_metrics = running_metrics(train_metrics, metric, sr.float(), hr.float())
+            train_metrics = running_metrics(train_metrics, metric, sr.float(), hr)
 
         if rank == 0:
             logger.log_dict({"Loss": train_loss / len(train_dl)}, epoch, "Train")
             logger.log_dict({k: v / len(train_dl) for k, v in train_metrics.items()}, epoch, "Train")
-            logger.log_images("Train", epoch, lr.float(), sr.float(), hr.float(), lq.float())
+            logger.log_images("Train", epoch, lr.float(), sr, hr, lq.float())
 
             print("Starting Evaluation ...")
